Refactor CoinFlip model and implement save/load functionality for new coinFlipQueue system.
Remove ChildPickingIndex. Replace functionality of queue by using an arrayList of integers which keeps track of the child indices and acts as a medium between the childList and coinFlipQueue. Allow for save/load of the separate queue to still be linked to singleton and not create copies of children when loading.